TempFileManager: adding mutex.

+++ Metalama.Backstage/Maintenance/TempFileManager.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) SharpCrafters s.r.o. See the LICENSE.md file in the root directory of this rep root for details.
 
 using Metalama.Backstage.Extensibility;
+using Metalama.Backstage.Utilities;
 using Newtonsoft.Json;
 using System;
 using System.IO;
@@ -26,21 +27,36 @@ internal class TempFileManager : ITempFileManager
             this._applicationInfoProvider.CurrentApplication.Version,
             guid?.ToString() ?? "" );
 
-        if ( !Directory.Exists( directory ) )
-        {
-            Directory.CreateDirectory( directory );
-        }
-
         var cleanUpFilePath = Path.Combine( directory, "cleanup.json" );
 
-        if ( !File.Exists( cleanUpFilePath ) )
+        if ( !Directory.Exists( directory ) || !File.Exists( cleanUpFilePath ) )
         {
-            var file = new CleanUpFile() { Strategy = cleanUpStrategy };
-            File.WriteAllText( cleanUpFilePath, JsonConvert.SerializeObject( file ) );
+            using ( MutexHelper.WithGlobalLock( directory ) )
+            {
+                if ( !Directory.Exists( directory ) || !File.Exists( cleanUpFilePath ) )
+                {
+                    if ( !Directory.Exists( directory ) )
+                    {
+                        Directory.CreateDirectory( directory );
+                    }
+
+                    if ( !File.Exists( cleanUpFilePath ) )
+                    {
+                        var file = new CleanUpFile() { Strategy = cleanUpStrategy };
+                        File.WriteAllText( cleanUpFilePath, JsonConvert.SerializeObject( file ) );
+
+                        return directory;
+                    }
+                }
+            }
         }
-        else if ( cleanUpStrategy == CleanUpStrategy.WhenUnused && File.GetLastAccessTime( cleanUpFilePath ) > DateTime.Now.AddDays( -1 ) )
+
+        if ( cleanUpStrategy == CleanUpStrategy.WhenUnused && File.GetLastAccessTime( cleanUpFilePath ) > DateTime.Now.AddDays( -1 ) )
         {
-            File.SetLastWriteTime( cleanUpFilePath, DateTime.Now );
+            using ( MutexHelper.WithGlobalLock( cleanUpFilePath ) )
+            {
+                File.SetLastWriteTime( cleanUpFilePath, DateTime.Now );
+            }
         }
 
         return directory;
